refactor: is_registered_childをChild/ChildrenModelに移動
